move copy_sign_and_docs decorator to helpers.py
